+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- ConsensusCommitPrologueV2 type transaction added to TxResult.
+
 ### Fixed
 
 ### Changed
@@ -16,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bumped libraries: dataclasses-json, httpx, websockets
 - Bumped dev libraries: mypy, pytest, pylint, sphinx_rtd_theme
 - [change](https://github.com/FrankC01/pysui/issues/175) Default 'compress_inputs' on SuiTransaction to True
+- Deprecating 'round' from RandomnessStateUpdate transaction type as of Sui 1.16.0
 
 ### Removed